# src/docs/swagger.yaml
basePath: /api
definitions:
  device.BindDeviceRequest:
    properties:
      device_id:
        type: string
    type: object
  device.BindDeviceResponse:
    properties:
      device_key:
        description: 设备长期密钥S（明文，仅下发一次）
        type: string
      message:
        description: 错误信息（失败时）
        type: string
      success:
        description: 是否成功
        type: boolean
      token:
        description: 30天有效期DeviceToken（成功时）
        type: string
    type: object
  device.RefreshTokenResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      token:
        type: string
    type: object
  device.UnbindDeviceRequest:
    properties:
      bind_key:
        type: string
      device_id:
        type: string
    required:
    - bind_key
    - device_id
    type: object
  device.UnbindDeviceResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: 怒喵 API 文档
  version: "1.0"
paths:
...
  /device/bind:
    post:
      consumes:
      - application/json
      description: 绑定设备到用户账户，生成设备密钥和token
      parameters:
      - description: Bearer <UserJWT>
        in: header
        name: Authorization
        required: true
        type: string
      - description: 设备绑定请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/device.BindDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功
          schema:
            $ref: '#/definitions/device.BindDeviceResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/device.BindDeviceResponse'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/device.BindDeviceResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/device.BindDeviceResponse'
      summary: 设备绑定
      tags:
      - Device
  /device/refresh_token:
    get:
      consumes:
      - application/json
      description: 刷新设备的JWT token，延长有效期
      parameters:
      - description: 设备ID
        in: query
        name: device_id
        required: true
        type: string
      - description: 当前设备token
        in: query
        name: token
        required: true
        type: string
      - description: 设备绑定密钥
        in: query
        name: bind_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            $ref: '#/definitions/device.RefreshTokenResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/device.RefreshTokenResponse'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/device.RefreshTokenResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/device.RefreshTokenResponse'
      summary: 刷新设备token
      tags:
      - Device
  /device/unbind:
    post:
      consumes:
      - application/json
      description: 解绑指定设备，将数据库中is_active字段设置为false
      parameters:
      - description: 设备解绑请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/device.UnbindDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 解绑成功
          schema:
            $ref: '#/definitions/device.UnbindDeviceResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/device.UnbindDeviceResponse'
        "404":
          description: 设备未找到
          schema:
            $ref: '#/definitions/device.UnbindDeviceResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/device.UnbindDeviceResponse'
      summary: 设备解绑
      tags:
      - Device
...
swagger: "2.0"